+++ cookbook/sdk/src/index.ts
@@ -1,19 +1,22 @@
-export { LeanCtxClient } from "./client.js";
-export { LeanCtxHttpError } from "./errors.js";
-export { toolResultToText } from "./toolText.js";
-export { runConformance } from "./conformance.js";
-
-export type {
-  CapabilitiesV1,
-  JsonObject,
-  JsonValue,
-  ContextEventV1,
-  ListToolsResponse,
-  ToolArguments,
-  ToolCallResponse,
-} from "./types.js";
 export type { LeanCtxClientOptions } from "./client.js";
+export { LeanCtxClient } from "./client.js";
 export type {
   ConformanceCheck,
   ConformanceScorecard,
 } from "./conformance.js";
+export {
+  COVERED_ROUTES,
+  runConformance,
+  SUPPORTED_HTTP_CONTRACT_VERSIONS,
+} from "./conformance.js";
+export { LeanCtxHttpError } from "./errors.js";
+export { toolResultToText } from "./toolText.js";
+export type {
+  CapabilitiesV1,
+  ContextEventV1,
+  JsonObject,
+  JsonValue,
+  ListToolsResponse,
+  ToolArguments,
+  ToolCallResponse,
+} from "./types.js";